+++ signer/src/signer/tools.c
@@ -210,10 +210,12 @@ tools_audit(zone_type* zone, engineconfig_type* config)
 
         se_log_debug("system call: %s", str);
         error = system(str);
-        if (!error) {
-            unlink(finalized);
+        if (finalized) {
+            if (!error) {
+                unlink(finalized);
+            }
+            se_free((void*)finalized);
         }
-        se_free((void*)finalized);
     }
     return error;
 }
